+++ github/src/lib.rs
@@ -1,0 +1,6 @@
+pub mod global;
+pub mod utils;
+pub mod fetch_github;
+pub mod models;
+
+pub mod routes;
