+++ github/src/routes.rs
@@ -1,0 +1,8 @@
+pub mod actions;
+pub mod auth;
+pub mod event;
+pub mod create_hook;
+pub mod hook_events;
+pub mod post;
+pub mod repos;
+pub mod revoke_hook;
