+++ github/src/routes/hook_events.rs
@@ -1,0 +1,11 @@
+use axum::{Json, response::IntoResponse};
+use serde_json::json;
+
+use crate::global::EVENTS;
+
+pub async fn hook_events() -> impl IntoResponse {
+    Json(json!({
+        "list": &*EVENTS
+    }))
+}
+
