+++ github/src/utils.rs
@@ -1,0 +1,37 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
+use rsa::{PaddingScheme, PublicKey};
+
+use crate::global::{CHACHA8RNG, PRIV_KEY, PUB_KEY};
+
+pub fn get_now() -> u64 {
+    let start = SystemTime::now();
+    let since_the_epoch = start
+        .duration_since(UNIX_EPOCH)
+        .expect("Time went backwards");
+    return since_the_epoch.as_secs();
+}
+
+pub fn encrypt(data: &str) -> String {
+    hex::encode(
+        PUB_KEY
+            .encrypt(
+                &mut CHACHA8RNG.clone(),
+                PaddingScheme::new_pkcs1v15_encrypt(),
+                data.as_bytes(),
+            )
+            .expect("failed to encrypt"),
+    )
+}
+
+pub fn decrypt(data: &str) -> String {
+    String::from_utf8(
+        PRIV_KEY
+            .decrypt(
+                PaddingScheme::new_pkcs1v15_encrypt(),
+                &hex::decode(data).unwrap(),
+            )
+            .expect("failed to decrypt"),
+    )
+    .unwrap()
+}
